Adjustments EngSwCIC#19 EngSwCIC#20 EngSwCIC#21 EngSwCIC#23
