+++ features/gerenciar_solicitacoes_credenciamento.feature
@@ -2,24 +2,24 @@
 #encoding: utf-8
 
 Funcionalidade: Gerenciar solicitações de credenciamento
-    Como um admnistrador do sistema
+    Como um admnistrador autenticado no sistema,
     Quero visualizar uma solicitação de credencimento em aberto
     Para decidir se vou aceitar ou recusar tal solicitação
 
     Contexto:
-        Dado que as seguintes solicitações estejam pendentes:
-        | title         | due_date    | activity_type_id              |
-        | Solicitação 1 | 02-Jan-2021 | Solicitação de credenciamento |
-        | Solicitação 2 | 02-Jan-2021 | Solicitação de credenciamento |
-        | Solicitação 3 | 02-Jan-2021 | Solicitação de credenciamento |
-        | Solicitação 4 | 02-Jan-2021 | Solicitação de credenciamento |
+        Dado que existam as seguintes solicitações:
+        | title         | activity_type_id              |
+        | Solicitação 1 | Solicitação de credenciamento |
+        | Solicitação 2 | Solicitação de credenciamento |
+        | Solicitação 3 | Solicitação de credenciamento |
+        | Solicitação 4 | Solicitação de credenciamento |
 
-        E que eu estou cadastrado e logado como "Gabriel", "gabriel@admin.com", "gabriel123", "administrator", "200000000"
-        E que eu estou na página de solicitações de credenciamento
+        E que eu esteja cadastrado e logado como "Gabriel", "gabriel@admin.com", "gabriel123", "administrator", "200000000"
+        E que eu esteja na página de solicitações de credenciamento
     
     Cenário: Aceitar uma solicitação de credenciamento
         Quando eu clico em "Solicitação 1"
-        Então eu devo estar na página da "Solicitação 1"
+        Então eu devo estar na página de "Solicitação 1"
         Quando eu clico em 'Aprovar'
         Então eu devo estar na página de solicitações de credenciamento
         Quando eu desmarco os seguintes estados: Rejeitadas, Reformulação
@@ -29,7 +29,7 @@ Funcionalidade: Gerenciar solicitações de credenciamento
     
     Cenário: Recusar uma solicitação de credenciamento
         Quando eu clico em "Solicitação 2"
-        Então eu devo estar na página da "Solicitação 2"
+        Então eu devo estar na página de "Solicitação 2"
         Quando eu clico em 'Rejeitar'
         Então eu devo estar na página de solicitações de credenciamento
         Quando eu desmarco os seguintes estados: Aprovadas, Reformulação
