+++ features/requisitos_necessarios.feature
@@ -1,22 +1,22 @@
 #language: pt
 #encoding: utf-8
 
-Funcionalidade: Disponibilizar os requisitos necessarios para credenciamento de professores
-  Como administrador autenticado no sistema,
-  Quero poder disponibilizar para os professores os requisitos necessários para o credenciamento
-  Para que eles possam dar procedimento ao credenciamento
+Funcionalidade: Disponibilizar os requisitos para o credenciamento
+  Como um administrador autenticado no sistema,
+  Para que os professores possam abrir solicitações de credenciamento
+  Quero poder disponibilizar os requisitos necessários para o credenciamento
 
   Contexto:
-    Dado que eu esteja logado como administrador de email "gp@admin.com" e senha "123"
-    E não existem requisitos selecionados na página principal
-    Quando o administrador clicou no link para alterar documentos necessários para credenciamento
+    Dado que eu esteja cadastrado e logado como "Preihs", "preihs@admin.com", "preihs123", "administrator", "200000000"
+    E que eu esteja na página de requisitos para o credenciamento
 
-    Cenário: Os campos puderam ser selecionados
-      Quando eu selecionar os campos
-      E eu clicar no botão atualizar requisitos
-      Então eu devo voltar para a página principal aonde aparece meus requisitos selecionados
-
-    Cenário: Não houveram mudanças feitas 
-      Quando eu não selecionar os campos
-      E eu clicar no botão atualizar requisitos
-      Então eu recebo uma mensagem dizendo que "não houveram mudanças"
+    Cenário: Modificar os requisitos necessários para o credenciamento
+      Quando eu preencho em 'Requisitos' com
+      """
+      - Formulário de solicitação preenchido e assinado pelo interessado*
+      - CV Lattes do coorientador* (apenas dados de publicação dos últimos 5 anos)
+      - Cópia do passaporte (apenas caso de orientador/coorientador estrangeiro)
+      """
+      E eu anexo o arquivo "Formulário de Credenciamento.doc" no campo 'Formulario'
+      E eu aperto 'Enviar'
+      Então eu devo ver "Requisitos atualizados com sucesso"
